feat: add database existence check in entry file

+++ jds/p8/bi/docker-entrypoint.sh
@@ -7,6 +7,11 @@ if [ -z "$DB_USER" ] || [ -z "$DB_PASSWORD" ] || [ -z "$DB_HOST" ] || [ -z "$DB_
     exit 1
 fi
 
+if ! mysql -u "$DB_USER" -P "$DB_PORT" -h "$DB_HOST" -e "CREATE DATABASE IF NOT EXISTS $DB_DATABASE;" 2>/dev/null; then
+    echo "Failed to create the database $DB_DATABASE. Exiting!"
+    exit 1
+fi
+
 if [ -d "$work_dir" ]; then
     envsubst < .env.template > .env
     envsubst < aerich_env.py.template > aerich_env.py
